+++ composenav/src/main/java/com/adamkobus/compose/navigation/model/NavHostComponents.kt
@@ -1,0 +1,9 @@
+package com.adamkobus.compose.navigation.model
+
+import com.adamkobus.compose.navigation.NavigationId
+
+internal class NavHostComponents(val navigationId: NavigationId) {
+
+    val navigationProcessor = NavigationProcessor(navigationId)
+    val destinationManager = NavStateManager(navigationId)
+}