+++ composenav/src/main/java/com/adamkobus/compose/navigation/model/NavigationConsumerImpl.kt
@@ -2,6 +2,7 @@ package com.adamkobus.compose.navigation.model
 
 import com.adamkobus.compose.navigation.ComposeNavigation
 import com.adamkobus.compose.navigation.NavigationConsumer
+import com.adamkobus.compose.navigation.action.DiscardReason
 import com.adamkobus.compose.navigation.action.NavAction
 import com.adamkobus.compose.navigation.action.NavActionWrapper
 import com.adamkobus.compose.navigation.action.NavigationResult
@@ -10,36 +11,54 @@ import kotlinx.coroutines.CompletableDeferred
 
 internal class NavigationConsumerImpl : NavigationConsumer {
 
-    private val navProcessor
-        get() = ComposeNavigation.getNavigationProcessor()
+    private val navProcessors
+        get() = ComposeNavigation.getAllNavProcessors()
 
     override fun offer(action: NavAction) {
-        navProcessor.postNavAction(action)
+        navProcessors.forEach { it.postNavAction(action) }
     }
 
-    override fun offer(action: NavActionWrapper) {
-        navProcessor.postNavAction(action.action)
+    override fun offer(wrapper: NavActionWrapper) {
+        navProcessors.forEach { it.postNavAction(wrapper.action) }
     }
 
     override fun offer(intent: NavIntent) {
-        navProcessor.postNavIntent(intent)
+        navProcessors.forEach { it.postNavIntent(intent) }
     }
 
     override suspend fun offerBlocking(action: NavAction): NavigationResult {
-        val deferred = CompletableDeferred<NavigationResult>()
-        navProcessor.postNavAction(action, onTaskCompleted = deferred)
-        return deferred.await()
+        val pairs = navProcessors.map {
+            Pair(it, CompletableDeferred<NavigationResult>())
+        }
+        pairs.forEach {
+            it.first.postNavAction(action, onTaskCompleted = it.second)
+        }
+        var result: NavigationResult = NavigationResult.Discarded(DiscardReason.NotDelivered)
+        pairs.forEach {
+            val innerResult = it.second.await()
+            if (result !is NavigationResult.Accepted) {
+                result = innerResult
+            }
+        }
+        return result
     }
 
-    override suspend fun offerBlocking(action: NavActionWrapper): NavigationResult {
-        val deferred = CompletableDeferred<NavigationResult>()
-        navProcessor.postNavAction(action.action, onTaskCompleted = deferred)
-        return deferred.await()
-    }
+    override suspend fun offerBlocking(wrapper: NavActionWrapper): NavigationResult = offerBlocking(wrapper.action)
 
     override suspend fun offerBlocking(intent: NavIntent): NavigationResult {
-        val deferred = CompletableDeferred<NavigationResult>()
-        navProcessor.postNavIntent(intent, onTaskCompleted = deferred)
-        return deferred.await()
+        val pairs = navProcessors.map {
+            Pair(it, CompletableDeferred<NavigationResult>())
+        }
+        pairs.forEach {
+            it.first.postNavIntent(intent, onTaskCompleted = it.second)
+        }
+        var result: NavigationResult = NavigationResult.Discarded(DiscardReason.NotDelivered)
+        pairs.forEach {
+            val innerResult = it.second.await()
+            if (result !is NavigationResult.Accepted) {
+                result = innerResult
+            }
+        }
+        return result
     }
 }